+++ docs/changelog.rst
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+- :feature:`6` Add :class:`structlog.processors.StackInfoRenderer` for adding stack information to log entries without involving exceptions.
+  Also added it to default processor chain.
 - :feature:`12` Allow optional positional arguments for :func:`structlog.get_logger` that are passed to logger factories.
   The standard library factory uses this for explicit logger naming.
 - :feature:`0` Add :class:`structlog.processors.ExceptionPrettyPrinter` for development and testing when multiline log entries aren't just acceptable but even helpful.